% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list_pdfs.R
\name{list_pdfs}
\alias{list_pdfs}
\title{List PDF files.}
\usage{
list_pdfs(loc)
}
\description{
\code{list_pdfs} a character string file name for PDF location and lists all PDFs.
MUST currently be using specified file structure.
}
