% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_pdf_to_df.R
\name{pdf_to_df}
\alias{pdf_to_df}
\title{Convert PDF to a dataframe.}
\usage{
pdf_to_df(pdf_file, type)
}
\arguments{
\item{pdf_file}{A PDF file}

\item{type}{The type of file - can be 'compstat' or 'pop'}
}
\value{
A data frame
}
\description{
\code{pdf_to_df()} takes a population or COMPSTAT report PDF and processes some of the data into
a dataframe.
}
