% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_pull_pdf.R
\name{pull_pdfs}
\alias{pull_pdfs}
\title{Pull PDFs from CDCR website.}
\usage{
pull_pdfs(source, dest)
}
\arguments{
\item{source}{A URL character string for population or COMPSTAT reports main page from CDCR}

\item{dest}{A file path character string for where to store the files on your machine}
}
\value{
Nothing, but PDFs are saved
}
\description{
\code{pull_pdfs()} take a URL from the CDCR website, and extracts all PDFs linked from that
page of the website. Currently only works for TPOP4 population reports and COMPSTAT reports.
}
