% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list_pdfs.R
\name{list_pdfs}
\alias{list_pdfs}
\title{List PDF files.}
\usage{
list_pdfs(loc)
}
\arguments{
\item{loc}{character string file name for PDF location and lists all PDFs.}
}
\value{
A vector of file names.
}
\description{
\code{list_pdfs} takes a folder location and lists all PDF files in that location.
}
